Add some control variables of makefile
Well... this should load the ~/.emacs.d/init.el which
will slow-down the compile-time but should work to enable
htmlize to do syntax highlight on source blocks.

The problem, however, seems that the theme is not working.
What I get for now is just a weird black/white syntax highlight.

There is this message as well on the end of the compilation:
Using schema /usr/share/emacs/25.3/etc/schema/xhtml.rnc